Move repositories to data module
